Improve error message when getting partition info.
Getting partition info happens when getting the event processor and
issues in the access key or policies come up as a result to this call.

After this commit, more information about the failure is tried to be
passed upstream so that the caller could get better details about the
error condition.

Fixes #20.